# 8300693: Lower the compile threshold and reduce the iterations of warmup loop in VarHandles tests