… test/jdk/java/lang/invoke/VarHandles/VarHandleTestByteArrayAsChar.java
… test/jdk/java/lang/invoke/VarHandles/VarHandleTestByteArrayAsDouble.java
… test/jdk/java/lang/invoke/VarHandles/VarHandleTestByteArrayAsFloat.java
… test/jdk/java/lang/invoke/VarHandles/VarHandleTestByteArrayAsInt.java
… test/jdk/java/lang/invoke/VarHandles/VarHandleTestByteArrayAsLong.java
… test/jdk/java/lang/invoke/VarHandles/VarHandleTestByteArrayAsShort.java
… test/jdk/java/lang/invoke/VarHandles/VarHandleTestMethodHandleAccessBoolean.java
… test/jdk/java/lang/invoke/VarHandles/VarHandleTestMethodHandleAccessByte.java
… test/jdk/java/lang/invoke/VarHandles/VarHandleTestMethodHandleAccessChar.java
… test/jdk/java/lang/invoke/VarHandles/VarHandleTestMethodHandleAccessDouble.java
… test/jdk/java/lang/invoke/VarHandles/VarHandleTestMethodHandleAccessFloat.java
… test/jdk/java/lang/invoke/VarHandles/VarHandleTestMethodHandleAccessInt.java
… test/jdk/java/lang/invoke/VarHandles/VarHandleTestMethodHandleAccessLong.java
… test/jdk/java/lang/invoke/VarHandles/VarHandleTestMethodHandleAccessShort.java
… test/jdk/java/lang/invoke/VarHandles/VarHandleTestMethodHandleAccessString.java
… test/jdk/java/lang/invoke/VarHandles/X-VarHandleTestAccess.java.template
… test/jdk/java/lang/invoke/VarHandles/X-VarHandleTestByteArrayView.java.template
… test/jdk/java/lang/invoke/VarHandles/X-VarHandleTestMethodHandleAccess.java.template